+++ rio_tiler/landsat8.py
@@ -40,7 +40,7 @@ LANDSAT_BUCKET = "s3://landsat-pds"
 LANDSAT_BANDS = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11"]
 
 # ref: https://docs.python.org/3/library/concurrent.futures.html#threadpoolexecutor
-MAX_THREADS = os.environ.get("MAX_THREADS", multiprocessing.cpu_count() * 5)
+MAX_THREADS = int(os.environ.get("MAX_THREADS", multiprocessing.cpu_count() * 5))
 
 
 def _landsat_get_mtl(sceneid):
